Moved trace.py to utils
This patch moves trace.py to the utils directory
