+++ aprsd/plugins/location.py
@@ -2,7 +2,8 @@ import logging
 import re
 import time
 
-from aprsd import plugin, plugin_utils, trace
+from aprsd import plugin, plugin_utils
+from aprsd.utils import trace
 
 
 LOG = logging.getLogger("APRSD")
